additional tests and refactor
