+++ scripts/test_results.json
@@ -1,1 +1,1 @@
-{"Tract Sum Total SUM_LBS_CHEMICAL": 0.1, "Tract Sum Total SUM_LBS_PRODUCT": 0.144, "School Districts Sum Total SUM_LBS_CHEMICAL": 0.1, "School Districts Sum Total SUM_LBS_PRODUCT": 0.144, "Zip Sum Total SUM_LBS_CHEMICAL": -1.914, "Zip Sum Total SUM_LBS_PRODUCT": -2.099, "Townships Total SUM_LBS_CHEMICAL": 0.1, "Townships Total SUM_LBS_PRODUCT": 0.144, "Section Total SUM_LBS_CHEMICAL": 0.1, "Section Total SUM_LBS_PRODUCT": 0.144}
+{"2022 County Totals :: SUM_LBS_CHEMICAL": "0.077%", "2022 County Totals :: SUM_LBS_PRODUCT": "0.531%", "2022 County correlation :: SUM_LBS_CHEMICAL r-squared": 100.0, "2022 County correlation :: SUM_LBS_PRODUCT r-squared": 99.957, "2020 County ag totals :: SUM_LBS_CHEMICAL": "0.06%", "2020 County ag totals :: SUM_LBS_PRODUCT": "0.084%", "2020 County ag correlation :: SUM_LBS_CHEMICAL r-squared": 100.0, "2020 County ag correlation :: SUM_LBS_PRODUCT r-squared": 100.0, "2022 County non-ag totals :: SUM_LBS_CHEMICAL": "0.109%", "2022 County non-ag totals :: SUM_LBS_PRODUCT": "0.664%", "2022 County non-ag correlation :: SUM_LBS_CHEMICAL r-squared": 99.999, "2022 County non-ag correlation :: SUM_LBS_PRODUCT r-squared": 98.286, "Tract Sum Total SUM_LBS_CHEMICAL": "0.1%", "Tract Sum Total SUM_LBS_PRODUCT": "0.144%", "School Districts Sum Total SUM_LBS_CHEMICAL": "0.1%", "School Districts Sum Total SUM_LBS_PRODUCT": "0.144%", "ZCTA Sum Total SUM_LBS_CHEMICAL": "-1.914%", "ZCTA Sum Total SUM_LBS_PRODUCT": "-2.099%", "Townships Total SUM_LBS_CHEMICAL": "0.1%", "Townships Total SUM_LBS_PRODUCT": "0.144%", "Section Total SUM_LBS_CHEMICAL": "0.1%", "Section Total SUM_LBS_PRODUCT": "0.144%", "Sections Sacramento alfalfa total POUNDS_CHEMICAL_APPLIED": "-0.0%", "Sections Sacramento alfalfa total POUNDS_PRODUCT_APPLIED": "-0.027%", "Sections Sacramento alfalfa correlation :: POUNDS_CHEMICAL_APPLIED r-squared": 100.0, "Sections Sacramento alfalfa correlation :: POUNDS_PRODUCT_APPLIED r-squared": 100.0}
